Refactor

+++ setup_project/setup_project.py
@@ -17,7 +17,7 @@ def helloworld(object):
     print("I am a {}.".format(object))
 
 
-def zip_push(sourcedir="project_name", zip_name="project_name"):
+def zip_push(sourcedir="project_name", file_name="project_name"):
     """
     Zips up the files in the source directory
 
@@ -26,13 +26,30 @@ def zip_push(sourcedir="project_name", zip_name="project_name"):
         zip_name (str): Name of the output zip file.
 
     Returns:
-        None
+        name (str): Name of the output zip file.
     """
+    if not file_name:
+        raise ValueError("Empty filename")
+    if not sourcedir:
+        raise ValueError("Directory doesn't exist")
     # Create zip file
-    shutil.make_archive(zip_name, "zip", sourcedir)
+    # Get full path of zip file created
+    name = ""
+    try:
+        name = shutil.make_archive(file_name, "zip", sourcedir)
+    except FileNotFoundError:
+        print("Source directory doesn't exist")
+
+    # get the start of the file name
+    start = 0 - (len(file_name + "zip") + 1)
+
+    # extract the filename from the full path
+    name = name[start:]
+
+    return name
 
 
-def zip_pop(file_name="project_name.zip"):
+def zip_pop(file_name="project_name.zip", ext_dir=None):
     """
     Unzips a zip file into the current directory
 
@@ -43,9 +60,24 @@ def zip_pop(file_name="project_name.zip"):
                          register_unpack_format().
 
     Returns:
-        None
+        bool: True if operation succeeded, False if not
     """
-    shutil.unpack_archive(filename=file_name)
+    # Check if file name has been passed
+    if not file_name:
+        raise ValueError("Empty filename")
+
+    # check if file is a zip file
+    if file_name[-4:] != ".zip":
+        raise ValueError("File extension should be zip")
+
+    # Check if file exists
+    if not os.path.exists(f"./{file_name}"):
+        raise FileNotFoundError(f"{file_name} doesn't exist")
+    else:
+        # unzip archive
+        shutil.unpack_archive(filename=file_name, extract_dir=ext_dir)
+        # return status
+        return True
 
 
 def del_file(file_name="project_name.zip"):
@@ -58,9 +90,17 @@ def del_file(file_name="project_name.zip"):
     Returns:
         None
     """
-    os.remove(file_name)
+    # Check if file name has been passed
+    if not file_name:
+        raise ValueError("Empty filename")
+
+    # Check if file exists
+    if not os.path.exists(f"./{file_name}"):
+        raise FileNotFoundError(f"{file_name} doesn't exist")
+    else:
+        os.remove(file_name)
+        return True
 
 
-if __name__ == "__main__":
-    # unzip the archive
-    zip_pop()
+if __name__ == "__main__":  # pragma: no cover
+    zip_push()
